Start a test for producing additional values in map.
Not sure what I want to do about the recursive aspect of this. Could
make the end user deal with it, as yeah, maybe they do want to
recursively be mapping. But wow is that uglier to deal with, and
recursively mapping yourself ought be pretty easy to implement.

So...  now need a state? To know produce is running?